refactor: move transformer/formatter utils to own file
