+++ src/formatters/index.ts
@@ -27,26 +27,4 @@ export const formatters: Record<string, Formatter> = {
   },
 };
 
-/**
- * Finds a formatter with the specified ID.
- */
-export function findFormatterById(id: number | string): Formatter | undefined {
-  if (typeof id === "string") {
-    id = parseInt(id, 10);
-  }
-
-  return Object.values(formatters).find((fm) => fm.id === id);
-}
-
-/**
- * Finds a formatter with the specified language.
- */
-export function findTransformerByLanguage(
-  language: Language
-): Formatter | undefined {
-  return Object.values(formatters).find((fm) => fm.language === language);
-}
-
-export * from "./css";
-export * from "./js";
-export * from "./jsx";
+export { formatCss, formatObject, formatProps };
