+++ src/utils/index.ts
@@ -2,5 +2,4 @@ export * from "./defaultHightlightProps";
 export * from "./exampleCode";
 export * from "./parser";
 export * from "./parseObj";
-export * from "./transformers";
 export * from "./usePrevious";
